+++ CHANGELOG.md
@@ -18,6 +18,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - Improve types for no labels
 - Faster stats gathering with lower memory overhead
 - Simplified number format logic
+- Fix memory leak in cluster.js by deleting all expired requests
 
 ### Added